Use Global Instant API
Cosmic time tries to be as atomic as possible. We want animations
to be frame perfect. The `Instant` is the value that is used to
determine percentage completion of the animation. Cosmic-time
has already had this concept with the `.start()` api. When
creating an animation, each chain is added to the timeline then
They are all started at the same instant with the `.start()`
method call. This prevents one widget's animation from being
slightly delayed from the others because of processing time
between adding each animation chain.

Though this logic was broken, because previously a widget would
generate it's own `Instant` when using `as_widget` to generate the
widget in the `view()`. But now the instant is routed through the
subscription and needs to be passed back to the timeline with the
`now()` method. This guarentees that animations will be timed
prefectly.