Replace keyframe Into<ExactSizeItterator> API
With recent changes to improve the `set_chain` and `start_at`
logic in `timeline.rs` the locgic wasn't iterating immedialy.
This meant that we had to collect the iterator into a vec in
`set_chain`. Which meant we were heap allocating twice for each
keyframe in the chain. By using `vec!` and
passign a vector to build `struct Chain`, that allocation can be
reused. Meaning only one heap allocation.